requirements.txt
google-generativeai
feedparser
edge-tts
pydub
python-dotenv
